fix(menu): circular dep on menu and sub
