+++ packages/elements/src/primitives/create-dom-collection/index.ts
@@ -1,3 +1,4 @@
 export * from "./create-dom-collection";
 export * from "./create-dom-collection-item";
+export * from "./dom-collection-context";
 export * from "./types";
